requirements.txt
admin-multiupload==1.10
certifi==2018.1.18
chardet==3.0.4
Django==2.0.3
django-appconf==1.0.2
django-embed-video==1.1.2
django-imagekit==4.0.2
idna==2.6
mysqlclient==1.3.12
pilkit==2.0
Pillow==5.0.0
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
